requirements.txt
cython
git+git://github.com/fantix/gevent.git#egg=gevent
pyzmq
